@@ -11,7 +11,7 @@
             "ES"
     ],
     "require": {
-        "php": ">=5.3.0"
+      "php": ">=5.4.0"
     },
     "require-dev": {
         "phpunit/phpunit": "3.7.38",
